requirements.txt
python-dotenv==1.0.0
cx-Oracle==8.3.0
pandas==2.0.3
scikit-learn==1.6.0
streamlit==1.40.2
plotly==5.18.0
